Move worklist state to own file
