+++ src/js/apps/patients/worklist/worklist_app.js
@@ -1,5 +1,5 @@
-import { clone, extend, reduce, keys, union } from 'underscore';
-import dayjs from 'dayjs';
+import { clone, extend, union } from 'underscore';
+
 import store from 'store';
 
 import Backbone from 'backbone';
@@ -9,138 +9,14 @@ import intl, { renderTemplate } from 'js/i18n';
 
 import App from 'js/base/app';
 
+import StateModel from './worklist_state';
+
 import FiltersApp from './filters_app';
 import BulkEditActionsApp from './sidebar/bulk-edit-actions_app';
 import BulkEditFlowsApp from './sidebar/bulk-edit-flows_app';
 
 import { ListView, SelectAllView, LayoutView, ListTitleView, TableHeaderView, SortDroplist, sortCreatedOptions, sortDueOptions, sortUpdateOptions } from 'js/views/patients/worklist/worklist_views';
 import { BulkEditButtonView, BulkEditFlowsSuccessTemplate, BulkEditActionsSuccessTemplate, BulkDeleteFlowsSuccessTemplate, BulkDeleteActionsSuccessTemplate } from 'js/views/patients/worklist/bulk-edit/bulk-edit_views';
-
-const StateModel = Backbone.Model.extend({
-  defaults() {
-    return {
-      actionsSortId: 'sortCreatedDesc',
-      flowsSortId: 'sortCreatedDesc',
-      filters: {
-        type: 'flows',
-        groupId: null,
-        clinicianId: this.currentClinician.id,
-        roleId: this.currentClinician.getRole().id,
-        selectedDate: null,
-        selectedMonth: null,
-        relativeDate: null,
-      },
-      selectedActions: {},
-      selectedFlows: {},
-    };
-  },
-  preinitialize() {
-    this.currentClinician = Radio.request('bootstrap', 'currentUser');
-    this.groups = this.currentClinician.getGroups();
-  },
-  initialize() {
-    this.on('change', this.onChange);
-  },
-  onChange() {
-    store.set(`${ this.id }_${ this.currentClinician.id }`, this.attributes);
-  },
-  getFilters() {
-    return clone(this.get('filters'));
-  },
-  getType() {
-    return this.getFilters().type;
-  },
-  isFlowType() {
-    return this.getType() === 'flows';
-  },
-  getSort() {
-    return this.get(`${ this.getType() }SortId`);
-  },
-  setSort(sortId) {
-    this.set(`${ this.getType() }SortId`, sortId);
-  },
-  getEntityFilter() {
-    const { groupId, clinicianId, roleId } = this.getFilters();
-    const group = groupId || this.groups.pluck('id').join(',');
-    const status = 'queued,started';
-
-    const filters = {
-      'owned-by': { status, group },
-      'shared-by': { status, group },
-      'new-past-day': {
-        created_since: dayjs().subtract(24, 'hours').format(),
-        status,
-        group,
-      },
-      'updated-past-three-days': {
-        updated_since: dayjs().startOf('day').subtract(3, 'days').format(),
-        status,
-        group,
-      },
-      'done-last-thirty-days': {
-        updated_since: dayjs().startOf('day').subtract(30, 'days').format(),
-        status: 'done',
-        group,
-      },
-    };
-
-    if (this.id === 'shared-by' || !clinicianId) {
-      filters[this.id].role = roleId;
-    } else {
-      filters[this.id].clinician = clinicianId;
-    }
-
-    return filters[this.id];
-  },
-  getSelectedList() {
-    return this.isFlowType() ? this.get('selectedFlows') : this.get('selectedActions');
-  },
-  toggleSelected(model, isSelected) {
-    const listName = this.isFlowType() ? 'selectedFlows' : 'selectedActions';
-    const list = clone(this.get(listName));
-
-    this.set(listName, extend(list, {
-      [model.id]: isSelected,
-    }));
-  },
-  isSelected(model) {
-    const list = this.getSelectedList();
-
-    return list[model.id];
-  },
-  getSelected(collection) {
-    const list = this.getSelectedList();
-    const collectionSelected = reduce(keys(list), (selected, item) => {
-      if (list[item] && collection.get(item)) {
-        selected.push({
-          id: item,
-        });
-      }
-
-      return selected;
-    }, []);
-
-    return Radio.request('entities', `${ this.getType() }:collection`, collectionSelected);
-  },
-  clearSelected() {
-    const listName = this.isFlowType() ? 'selectedFlows' : 'selectedActions';
-
-    this.set(listName, {});
-    this.trigger('select:none');
-  },
-  selectAll(collection) {
-    const listName = this.isFlowType() ? 'selectedFlows' : 'selectedActions';
-
-    const list = collection.reduce((selected, model) => {
-      selected[model.id] = true;
-      return selected;
-    }, {});
-
-    this.set(listName, list);
-
-    this.trigger('select:all');
-  },
-});
 
 export default App.extend({
   StateModel,
